Create kwatrofound.py
a.	There are two opponents in the game (in this case, a human player and the computer). 
b.	The game is played on two identical boards (10-by-10 tables of squares) on which each opponent places his or her ships, not seen by the opponent.
c.	Each player has five ships, each of which occupies a certain number of squares on the board: 
                   i.	a destroyer (2 squares), 
                   ii.	a submarine (3 squares), 
                   iii.	a cruiser (3 squares), 
                   iv.	a battleship (4 squares), and 
                   v.	an aircraft carrier (5 squares). 
d.	Each ship is placed either horizontally or vertically, with no two ships touching each other. 
e.	The game is played by the opponents taking turns “shooting” at each other’s ships. 
f.	A result of every shot is displayed as either a hit or a miss. In case of a hit, the player gets to go again and keeps playing until this player misses. 
g.	The goal is to sink all the opponent’s ships before the opponent succeeds in doing it first. (To sink a ship, all squares occupied by the ship must be hit.)